requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
Django==3.1.4
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==3.3.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pkg-resources==0.0.0
pluggy==0.13.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.4
rope==0.18.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==4.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
